corrigido erro ao tentar renomear a pasta dados a cada inicialização

+++ dados.py
@@ -46,9 +46,10 @@ dados_config = {
 # Diretórios base
 dados_dir = "dados"
 
-if os.path.exists("Dados") and os.path.isfile("CopiarXML3.jar"):
-    os.rename("Dados", "Dados.old")
-    os.rename("Imagens", "imagens")
+if not os.path.exists("Dados.old"):
+    if os.path.exists("Dados") and os.path.isfile("CopiarXML3.jar"):
+        os.rename("Dados", "Dados.old")
+        os.rename("Imagens", "imagens")
 
 if not os.path.exists(dados_dir):
     os.makedirs(dados_dir)
